Migrate to new impl.
