+++ crates/action-graph/src/action_graph_builder.rs
@@ -1,14 +1,17 @@
 use crate::action_graph::ActionGraph;
+use miette::IntoDiagnostic;
 use moon_action::{
-    ActionNode, InstallProjectDepsNode, InstallWorkspaceDepsNode, RunTaskNode,
-    SetupToolchainLegacyNode, SetupToolchainNode, SyncProjectNode,
+    ActionNode, InstallDependenciesNode, InstallProjectDepsNode, InstallWorkspaceDepsNode,
+    RunTaskNode, SetupEnvironmentNode, SetupToolchainLegacyNode, SetupToolchainNode,
+    SyncProjectNode,
 };
 use moon_action_context::{ActionContext, TargetState};
 use moon_affected::{AffectedTracker, DownstreamScope, UpstreamScope};
 use moon_app_context::AppContext;
-use moon_common::path::WorkspaceRelativePathBuf;
+use moon_common::path::{PathExt, WorkspaceRelativePathBuf, is_root_level_source};
 use moon_common::{Id, color};
 use moon_config::{PipelineActionSwitch, TaskDependencyConfig};
+use moon_pdk_api::LocateDependenciesRootInput;
 use moon_platform::{PlatformManager, Runtime, ToolchainSpec};
 use moon_project::Project;
 use moon_query::{Criteria, build_query};
@@ -16,33 +19,24 @@ use moon_task::{Target, TargetError, TargetLocator, TargetScope, Task};
 use moon_task_args::parse_task_args;
 use moon_workspace_graph::{GraphConnections, WorkspaceGraph, tasks::TaskGraphError};
 use petgraph::prelude::*;
+use petgraph::visit::IntoNodeReferences;
 use rustc_hash::{FxHashMap, FxHashSet};
+use starbase_utils::glob::GlobSet;
 use std::mem;
 use std::sync::Arc;
 use tracing::{debug, instrument, trace};
 
 #[derive(Default)]
 pub struct RunRequirements {
-    pub ci: bool,         // Are we in a CI environment
-    pub ci_check: bool,   // Check the `runInCI` option
-    pub dependents: bool, // Run dependent tasks as well
-    pub interactive: bool,
-    pub skip_affected: bool, // Temporary until we support task dependents properly
-    pub target_locators: FxHashSet<TargetLocator>,
-}
-
-impl RunRequirements {
-    pub fn has_target(&self, target: &Target) -> bool {
-        self.target_locators.iter().any(|loc| match loc {
-            TargetLocator::GlobMatch { .. } => false,
-            TargetLocator::Qualified(other) => other == target,
-            TargetLocator::TaskFromWorkingDir(task_id) => task_id == &target.task_id,
-        })
-    }
+    pub ci: bool,          // Are we in a CI environment
+    pub ci_check: bool,    // Check the `runInCI` option
+    pub dependents: bool,  // Run dependent tasks as well
+    pub interactive: bool, // Entire pipeline is interactive
 }
 
 pub struct ActionGraphBuilderOptions {
     pub install_dependencies: PipelineActionSwitch,
+    pub setup_environment: PipelineActionSwitch, // TODO
     pub setup_toolchains: PipelineActionSwitch,
     pub sync_projects: PipelineActionSwitch,
     pub sync_project_dependencies: bool,
@@ -59,6 +53,7 @@ impl ActionGraphBuilderOptions {
     pub fn new(state: bool) -> Self {
         Self {
             install_dependencies: state.into(),
+            setup_environment: state.into(),
             setup_toolchains: state.into(),
             sync_projects: state.into(),
             sync_project_dependencies: state,
@@ -67,41 +62,35 @@ impl ActionGraphBuilderOptions {
     }
 }
 
-pub struct ActionGraphBuilder<'app> {
-    all_query: Option<Criteria<'app>>,
+// sync_workspace
+//   - change workspace/root files
+//   - change toolchain files
+// sync_project
+//   - change project files/manifests
+//   - change toolchain files
+// install_deps:
+// setup_toolchain:
+// run_task:
+pub struct ActionGraphBuilder<'query> {
+    all_query: Option<Criteria<'query>>,
     app_context: Arc<AppContext>,
     graph: DiGraph<ActionNode, ()>,
-    indices: FxHashMap<ActionNode, NodeIndex>,
     options: ActionGraphBuilderOptions,
-    platform_manager: &'app PlatformManager,
+    platform_manager: Option<PlatformManager>,
     workspace_graph: Arc<WorkspaceGraph>,
 
-    // Affected states
+    // Affected tracking
     affected: Option<AffectedTracker>,
     touched_files: Option<FxHashSet<WorkspaceRelativePathBuf>>,
 
     // Target tracking
-    initial_targets: FxHashSet<Target>,
+    // initial_targets: FxHashSet<Target>,
     passthrough_targets: FxHashSet<Target>,
     primary_targets: FxHashSet<Target>,
 }
 
-impl<'app> ActionGraphBuilder<'app> {
+impl<'query> ActionGraphBuilder<'query> {
     pub fn new(
-        app_context: Arc<AppContext>,
-        workspace_graph: Arc<WorkspaceGraph>,
-        options: ActionGraphBuilderOptions,
-    ) -> miette::Result<Self> {
-        ActionGraphBuilder::with_platforms(
-            PlatformManager::read(),
-            app_context,
-            workspace_graph,
-            options,
-        )
-    }
-
-    pub fn with_platforms(
-        platform_manager: &'app PlatformManager,
         app_context: Arc<AppContext>,
         workspace_graph: Arc<WorkspaceGraph>,
         options: ActionGraphBuilderOptions,
@@ -109,34 +98,30 @@ impl<'app> ActionGraphBuilder<'app> {
         debug!("Building action graph");
 
         Ok(ActionGraphBuilder {
-            all_query: None,
             affected: None,
+            all_query: None,
             app_context,
             graph: DiGraph::new(),
-            indices: FxHashMap::default(),
-            initial_targets: FxHashSet::default(),
+            // initial_targets: FxHashSet::default(),
             options,
             passthrough_targets: FxHashSet::default(),
-            platform_manager,
+            platform_manager: None,
             primary_targets: FxHashSet::default(),
-            workspace_graph,
             touched_files: None,
+            workspace_graph,
         })
     }
 
-    pub fn build(self) -> ActionGraph {
-        ActionGraph::new(self.graph)
-    }
-
-    pub fn build_context(&mut self) -> ActionContext {
+    pub fn build(mut self) -> (ActionContext, ActionGraph) {
         let mut context = ActionContext {
             affected: self.affected.take().map(|affected| affected.build()),
             ..ActionContext::default()
         };
 
-        if !self.initial_targets.is_empty() {
-            context.initial_targets = mem::take(&mut self.initial_targets);
-        }
+        // TODO
+        // if !self.initial_targets.is_empty() {
+        //     context.initial_targets = mem::take(&mut self.initial_targets);
+        // }
 
         if !self.passthrough_targets.is_empty() {
             for target in mem::take(&mut self.passthrough_targets) {
@@ -152,15 +137,21 @@ impl<'app> ActionGraphBuilder<'app> {
             context.touched_files = files.to_owned();
         }
 
-        context
+        (context, ActionGraph::new(self.graph))
     }
 
-    pub fn get_index_from_node(&self, node: &ActionNode) -> Option<&NodeIndex> {
-        self.indices.get(node)
-    }
-
-    pub fn get_runtime(&self, project: &Project, toolchain: &Id, allow_override: bool) -> Runtime {
-        if let Ok(platform) = self.platform_manager.get_by_toolchain(toolchain) {
+    pub fn get_runtime(
+        &self,
+        project: &Project,
+        toolchain_id: &Id,
+        allow_override: bool,
+    ) -> Runtime {
+        if let Ok(platform) = self
+            .platform_manager
+            .as_ref()
+            .unwrap()
+            .get_by_toolchain(toolchain_id)
+        {
             return platform.get_runtime_from_config(if allow_override {
                 Some(&project.config)
             } else {
@@ -174,72 +165,55 @@ impl<'app> ActionGraphBuilder<'app> {
     pub fn get_spec(
         &self,
         project: &Project,
-        toolchain: &Id,
+        toolchain_id: &Id,
         allow_override: bool,
     ) -> Option<ToolchainSpec> {
-        if let Some(config) = project.config.toolchain.plugins.get(toolchain) {
+        if let Some(config) = project.config.toolchain.plugins.get(toolchain_id) {
             if !config.is_enabled() {
                 return None;
             }
 
-            if let Some(version) = config.get_version() {
-                if allow_override {
+            if allow_override {
+                if let Some(version) = config.get_version() {
                     return Some(ToolchainSpec::new_override(
-                        toolchain.to_owned(),
+                        toolchain_id.to_owned(),
                         version.to_owned(),
                     ));
                 }
             }
         }
 
-        if let Some(config) = self.app_context.toolchain_config.plugins.get(toolchain) {
-            return Some(if let Some(version) = &config.version {
-                ToolchainSpec::new(toolchain.to_owned(), version.to_owned())
-            } else {
-                ToolchainSpec::new_global(toolchain.to_owned())
+        if let Some(config) = self.app_context.toolchain_config.plugins.get(toolchain_id) {
+            return Some(match &config.version {
+                Some(version) => ToolchainSpec::new(toolchain_id.to_owned(), version.to_owned()),
+                None => ToolchainSpec::new_global(toolchain_id.to_owned()),
             });
         }
 
         None
     }
 
-    pub fn set_affected(&mut self) {
-        let Some(touched_files) = &self.touched_files else {
-            return;
-        };
-
+    pub fn set_affected(&mut self) -> miette::Result<()> {
         if self.affected.is_none() {
             self.affected = Some(AffectedTracker::new(
                 Arc::clone(&self.workspace_graph),
-                touched_files.to_owned(),
+                self.touched_files
+                    .as_ref()
+                    .expect("Touched files are required for affected tracking.")
+                    .to_owned(),
             ));
         }
-    }
-
-    pub fn set_affected_scopes(
-        &mut self,
-        upstream: UpstreamScope,
-        downstream: DownstreamScope,
-    ) -> miette::Result<()> {
-        // If we require dependents, then we must load all projects into the
-        // graph so that the edges are created!
-        if downstream != DownstreamScope::None {
-            debug!("Force loading all projects and tasks to determine relationships");
-
-            self.workspace_graph.get_projects()?;
-            self.workspace_graph.get_tasks_with_internal()?;
-        }
-
-        self.set_affected();
-        self.affected
-            .as_mut()
-            .unwrap()
-            .with_scopes(upstream, downstream);
 
         Ok(())
     }
 
-    pub fn set_query(&mut self, input: &'app str) -> miette::Result<()> {
+    pub fn set_platform_manager(&mut self, manager: PlatformManager) -> miette::Result<()> {
+        self.platform_manager = Some(manager);
+
+        Ok(())
+    }
+
+    pub fn set_query(&mut self, input: &'query str) -> miette::Result<()> {
         self.all_query = Some(build_query(input)?);
 
         Ok(())
@@ -254,311 +228,372 @@ impl<'app> ActionGraphBuilder<'app> {
         Ok(())
     }
 
-    // ACTIONS
+    pub fn track_affected(
+        &mut self,
+        upstream: UpstreamScope,
+        downstream: DownstreamScope,
+        ci_check: bool,
+    ) -> miette::Result<()> {
+        // If we require dependents, then we must load all projects into the
+        // graph so that the edges are created!
+        if downstream != DownstreamScope::None {
+            debug!("Force loading all projects and tasks to determine relationships");
+
+            self.workspace_graph.get_projects()?;
+            self.workspace_graph.get_tasks_with_internal()?;
+        }
+
+        self.set_affected()?;
+
+        if let Some(affected) = self.affected.as_mut() {
+            affected.set_ci_check(ci_check);
+            affected.with_scopes(upstream, downstream);
+            affected.track_projects()?;
+            affected.track_tasks()?;
+        }
+
+        Ok(())
+    }
 
     #[instrument(skip_all)]
-    pub fn install_deps(
+    pub async fn install_dependencies_legacy(
         &mut self,
+        runtime: &Runtime,
         project: &Project,
-        task: Option<&Task>,
+        has_bun_and_node: bool,
     ) -> miette::Result<Option<NodeIndex>> {
-        let mut in_project = false;
-        let toolchains = task
-            .map(|t| &t.toolchains)
-            .unwrap_or_else(|| &project.toolchains);
-        let mut primary_toolchain = toolchains[0].to_owned();
-        let mut packages_root = WorkspaceRelativePathBuf::default();
+        let setup_toolchain_index = self.setup_toolchain_legacy(runtime).await?;
+
+        if !self
+            .options
+            .install_dependencies
+            .is_enabled(&runtime.toolchain)
+            || runtime.is_system()
+        {
+            return Ok(setup_toolchain_index);
+        }
 
         // If Bun and Node.js are enabled, they will both attempt to install
         // dependencies in the target root. We need to avoid this problem,
         // so always prefer Node.js instead. Revisit in the future.
-        if primary_toolchain == "bun"
-            && self
-                .platform_manager
-                .enabled()
-                .any(|enabled_platform| enabled_platform == "node")
-        {
+        if has_bun_and_node && runtime.toolchain == "bun" {
             debug!(
                 "Already installing dependencies with node, skipping a conflicting install from bun"
             );
 
-            primary_toolchain = Id::raw("node")
+            return Ok(setup_toolchain_index);
         }
+
+        let platform = self
+            .platform_manager
+            .as_ref()
+            .unwrap()
+            .get_by_toolchain(&runtime.toolchain)?;
+
+        let packages_root = platform.find_dependency_workspace_root(project.source.as_str())?;
+        let mut in_project = false;
 
         // If project is NOT in the package manager workspace, then we should
         // install dependencies in the project, not the workspace root.
-        if let Ok(platform) = self.platform_manager.get_by_toolchain(&primary_toolchain) {
-            packages_root = platform.find_dependency_workspace_root(project.source.as_str())?;
+        if !platform.is_project_in_dependency_workspace(&packages_root, project.source.as_str())? {
+            in_project = true;
 
-            if !platform
-                .is_project_in_dependency_workspace(&packages_root, project.source.as_str())?
-            {
-                in_project = true;
+            debug!(
+                "Project {} is not within the dependency manager workspace, dependencies will be installed within the project instead of the root",
+                color::id(&project.id),
+            );
+        }
 
-                debug!(
-                    "Project {} is not within the dependency manager workspace, dependencies will be installed within the project instead of the root",
-                    color::id(&project.id),
+        let index = self.insert_node(if in_project {
+            ActionNode::install_project_deps(InstallProjectDepsNode {
+                project_id: project.id.to_owned(),
+                runtime: runtime.to_owned(),
+            })
+        } else {
+            ActionNode::install_workspace_deps(InstallWorkspaceDepsNode {
+                runtime: runtime.to_owned(),
+                root: packages_root,
+            })
+        });
+
+        // Before we install deps, we must ensure the language has been installed
+        self.link_requirements(index, vec![setup_toolchain_index]);
+
+        Ok(Some(index))
+    }
+
+    #[instrument(skip_all)]
+    pub async fn install_dependencies(
+        &mut self,
+        spec: &ToolchainSpec,
+        project: &Project,
+    ) -> miette::Result<Option<NodeIndex>> {
+        let setup_toolchain_index = self.setup_toolchain(spec).await?;
+
+        // Explicitly disabled
+        if !self.options.install_dependencies.is_enabled(&spec.id) || spec.is_system() {
+            return Ok(setup_toolchain_index);
+        }
+
+        let registry = &self.app_context.toolchain_registry;
+        let toolchain = registry.load(&spec.id).await?;
+
+        // Toolchain does not support this action, so skip and fall through
+        if !toolchain.supports_tier_2().await {
+            return Ok(setup_toolchain_index);
+        }
+
+        let output = toolchain
+            .locate_dependencies_root(LocateDependenciesRootInput {
+                context: registry.create_context(),
+                starting_dir: toolchain.to_virtual_path(&project.root),
+            })
+            .await?;
+
+        // Only insert this action if a root was located
+        if let Some(root) = output.root {
+            let abs_root = toolchain.from_virtual_path(root.any_path());
+            let rel_root = abs_root
+                .relative_to(&self.app_context.workspace_root)
+                .into_diagnostic()?;
+
+            // Determine if we're in the dependencies workspace
+            let in_project = project.root == abs_root;
+            let in_workspace = if let Some(globs) = output.members {
+                if in_project {
+                    true // Root always in the workspace
+                } else {
+                    GlobSet::new(&globs)?.matches(project.source.as_str())
+                }
+            } else {
+                true
+            };
+
+            // If not in the dependencies workspace (if there is one),
+            // or is a stand-alone project with its own lockfile,
+            // we must extract the project ID and source (root)
+            let (project_id, root) =
+                if !in_workspace || in_project && !is_root_level_source(&project.source) {
+                    (Some(project.id.clone()), project.source.clone())
+                } else {
+                    (None, rel_root)
+                };
+
+            let setup_env = ActionNode::setup_environment(SetupEnvironmentNode {
+                project_id: project_id.clone(),
+                root: root.clone(),
+                toolchain_id: spec.id.clone(),
+            });
+
+            let install_deps = ActionNode::install_dependencies(InstallDependenciesNode {
+                project_id,
+                root,
+                toolchain_id: spec.id.clone(),
+            });
+
+            // We need to conditionally create nodes and edges based on what
+            // APIs have been implemented by the plugin
+            let has_install_deps = toolchain.has_func("install_dependencies").await;
+            let has_setup_env = toolchain.has_func("setup_environment").await;
+
+            let index = match (has_install_deps, has_setup_env) {
+                (true, true) => {
+                    let setup_env_index = self.insert_node(setup_env);
+                    let install_deps_index = self.insert_node(install_deps);
+
+                    self.link_requirements(install_deps_index, vec![Some(setup_env_index)]);
+                    self.link_requirements(setup_env_index, vec![setup_toolchain_index]);
+
+                    Some(install_deps_index)
+                }
+                (true, false) => {
+                    let install_deps_index = self.insert_node(install_deps);
+
+                    self.link_requirements(install_deps_index, vec![setup_toolchain_index]);
+
+                    Some(install_deps_index)
+                }
+                (false, true) => {
+                    let setup_env_index = self.insert_node(setup_env);
+
+                    self.link_requirements(setup_env_index, vec![setup_toolchain_index]);
+
+                    Some(setup_env_index)
+                }
+                (false, false) => setup_toolchain_index,
+            };
+
+            return Ok(index);
+        }
+
+        Ok(setup_toolchain_index)
+    }
+
+    #[instrument(skip_all)]
+    pub async fn install_dependencies_by_project(
+        &mut self,
+        project: &Project,
+    ) -> miette::Result<Vec<Option<NodeIndex>>> {
+        self.install_dependencies_by_toolchains(project, &project.toolchains)
+            .await
+    }
+
+    #[instrument(skip_all)]
+    pub async fn install_dependencies_by_toolchains(
+        &mut self,
+        project: &Project,
+        toolchains: &[Id],
+    ) -> miette::Result<Vec<Option<NodeIndex>>> {
+        let mut indexes = vec![];
+        let has_bun_and_node =
+            toolchains.iter().any(|tc| tc == "node") && toolchains.iter().any(|tc| tc == "bun");
+
+        for toolchain_id in toolchains {
+            if self.app_context.toolchain_config.is_plugin(toolchain_id) {
+                if let Some(spec) = self.get_spec(project, toolchain_id, true) {
+                    indexes.push(self.install_dependencies(&spec, project).await?);
+                }
+            } else {
+                let runtime = self.get_runtime(project, toolchain_id, true);
+
+                indexes.push(
+                    self.install_dependencies_legacy(&runtime, project, has_bun_and_node)
+                        .await?,
                 );
             }
         }
 
-        let node = if in_project {
-            ActionNode::install_project_deps(InstallProjectDepsNode {
-                project_id: project.id.to_owned(),
-                runtime: self.get_runtime(project, &primary_toolchain, true),
-            })
-        } else {
-            ActionNode::install_workspace_deps(InstallWorkspaceDepsNode {
-                runtime: self.get_runtime(project, &primary_toolchain, false),
-                root: packages_root,
-            })
-        };
+        Ok(indexes)
+    }
 
-        let mut edges = vec![];
+    pub async fn run_task(
+        &mut self,
+        task: &Task,
+        reqs: &RunRequirements,
+    ) -> miette::Result<Option<NodeIndex>> {
+        if let Some(index) = self.internal_run_task(task, reqs, None).await? {
+            // Only track primary targets at the top-level run methods,
+            // as these are explicitly called by pipeline consumers!
+            self.primary_targets.insert(task.target.clone());
 
-        // Before we install deps, we must ensure the language has been installed
-        if let Some(spec) = self.get_spec(project, &primary_toolchain, in_project) {
-            if let Some(edge) = self.setup_toolchain_plugin(&spec) {
-                edges.push(edge);
-            }
-        } else {
-            let project_runtime = self.get_runtime(project, &primary_toolchain, true);
+            return Ok(Some(index));
+        }
 
-            if let Some(edge) = self.setup_toolchain(&project_runtime) {
-                edges.push(edge);
-            }
-        };
+        Ok(None)
+    }
 
-        // If installing dependencies is disabled, we still need to ensure the toolchain
-        // has been setup, and indirectly, the sync workspace action
-        if !self
-            .options
-            .install_dependencies
-            .is_enabled(&primary_toolchain)
-            || node.get_runtime().is_system()
+    #[cfg(debug_assertions)]
+    pub async fn run_task_with_config(
+        &mut self,
+        task: &Task,
+        reqs: &RunRequirements,
+        config: &TaskDependencyConfig,
+    ) -> miette::Result<Option<NodeIndex>> {
+        self.internal_run_task(task, reqs, Some(config)).await
+    }
+
+    pub async fn run_task_by_target<T: AsRef<Target>>(
+        &mut self,
+        target: T,
+        reqs: &RunRequirements,
+    ) -> miette::Result<FxHashSet<NodeIndex>> {
+        let target = target.as_ref();
+        let mut indexes = FxHashSet::default();
+
+        for task in self
+            .internal_resolve_tasks_from_target(target, false)
+            .await?
         {
-            return Ok(edges.first().cloned());
-        }
-
-        let index = match self.get_index_from_node(&node) {
-            Some(i) => *i,
-            None => self.insert_node(node),
-        };
-
-        self.link_requirements(index, edges);
-
-        Ok(Some(index))
-    }
-
-    pub fn run_task(
-        &mut self,
-        project: &Project,
-        task: &Task,
-        reqs: &RunRequirements,
-    ) -> miette::Result<Option<NodeIndex>> {
-        self.run_task_with_config(project, task, reqs, None)
-    }
-
-    #[instrument(skip_all)]
-    pub fn run_task_with_config(
-        &mut self,
-        project: &Project,
-        task: &Task,
-        reqs: &RunRequirements,
-        config: Option<&TaskDependencyConfig>,
-    ) -> miette::Result<Option<NodeIndex>> {
-        // Create a new requirements object as we don't want our dependencies/
-        // dependents to check for affected or run their own dependents!
-        let child_reqs = RunRequirements {
-            ci: reqs.ci,
-            ci_check: reqs.ci_check,
-            dependents: false,
-            interactive: reqs.interactive,
-            skip_affected: true,
-            ..Default::default()
-        };
-
-        // Only apply checks when requested. This applies to `moon ci`,
-        // but not `moon run`, since the latter should be able to
-        // manually run local tasks in CI (deploys, etc).
-        if reqs.ci && reqs.ci_check && !task.should_run_in_ci() {
-            self.passthrough_targets.insert(task.target.clone());
-
-            debug!(
-                task_target = task.target.as_str(),
-                "Not running task {} because {} is false",
-                color::label(&task.target.id),
-                color::property("runInCI"),
-            );
-
-            // Dependents may still want to run though,
-            // but only if this task was affected
-            if reqs.dependents {
-                if let Some(affected) = &mut self.affected {
-                    trace!(
-                        task_target = task.target.as_str(),
-                        "But will run all dependent tasks if affected"
-                    );
-
-                    if affected.is_task_marked(task) {
-                        self.run_task_dependents(task, &child_reqs)?;
-                    }
-                }
-            }
-
-            return Ok(None);
-        }
-
-        // These tasks shouldn't actually run, so filter them out
-        if self.passthrough_targets.contains(&task.target) {
-            trace!(
-                task_target = task.target.as_str(),
-                "Not adding task {} to graph because it has been marked as passthrough",
-                color::label(&task.target.id),
-            );
-
-            return Ok(None);
-        }
-
-        let mut args = vec![];
-        let mut env = FxHashMap::default();
-
-        if let Some(config) = config {
-            args.extend(parse_task_args(&config.args)?);
-            env.extend(config.env.clone());
-        }
-
-        let node = ActionNode::run_task(RunTaskNode {
-            args,
-            env,
-            interactive: task.is_interactive() || reqs.interactive,
-            persistent: task.is_persistent(),
-            priority: task.options.priority.get_level(),
-            target: task.target.to_owned(),
-            id: None,
-        });
-
-        if let Some(index) = self.get_index_from_node(&node) {
-            return Ok(Some(*index));
-        }
-
-        // Compare against touched files if provided
-        if let Some(affected) = &mut self.affected {
-            if !reqs.skip_affected && !affected.is_task_marked(task) {
-                return Ok(None);
+            if let Some(index) = self.run_task(&task, reqs).await? {
+                indexes.insert(index);
             }
         }
 
-        // We should install deps & sync projects *before* running targets
-        let mut edges = vec![];
-
-        if let Some(install_deps_index) = self.install_deps(project, Some(task))? {
-            edges.push(install_deps_index);
-        }
-
-        if let Some(sync_project_index) = self.sync_project(project)? {
-            edges.push(sync_project_index);
-        }
-
-        // Insert the node and create edges
-        let index = self.insert_node(node);
-
-        // And we also need to create edges for task dependencies
-        if !task.deps.is_empty() {
-            trace!(
-                task_target = task.target.as_str(),
-                dep_targets = ?task.deps.iter().map(|d| d.target.as_str()).collect::<Vec<_>>(),
-                "Linking dependencies for task",
-            );
-
-            edges.extend(self.run_task_dependencies(task, &child_reqs)?);
-        }
-
-        self.link_requirements(index, edges);
-
-        // And possibly dependents
-        if reqs.dependents {
-            self.run_task_dependents(task, &child_reqs)?;
-        }
-
-        Ok(Some(index))
+        Ok(indexes)
     }
 
-    // We don't pass touched files to dependencies, because if the parent
-    // task is affected/going to run, then so should all of these!
-    #[instrument(skip_all)]
-    pub fn run_task_dependencies(
+    pub async fn run_task_by_target_locator<T: AsRef<TargetLocator>>(
+        &mut self,
+        locator: T,
+        reqs: &RunRequirements,
+    ) -> miette::Result<FxHashSet<NodeIndex>> {
+        let locator = locator.as_ref();
+        let mut indexes = FxHashSet::default();
+
+        for task in self
+            .internal_resolve_tasks_from_target_locator(locator, false)
+            .await?
+        {
+            if let Some(index) = self.run_task(&task, reqs).await? {
+                indexes.insert(index);
+            }
+        }
+
+        Ok(indexes)
+    }
+
+    pub async fn run_task_dependencies(
         &mut self,
         task: &Task,
         reqs: &RunRequirements,
-    ) -> miette::Result<Vec<NodeIndex>> {
+    ) -> miette::Result<Vec<Option<NodeIndex>>> {
         let parallel = task.options.run_deps_in_parallel;
-        let mut indices = vec![];
-        let mut previous_target_index = None;
+        let mut indexes: Vec<Option<NodeIndex>> = vec![];
+        let mut previous_target_index: Option<NodeIndex> = None;
 
         for dep in &task.deps {
-            let (_, dep_indices) =
-                self.run_task_by_target_with_config(&dep.target, reqs, Some(dep))?;
-
-            for dep_index in dep_indices {
-                // When parallel, parent depends on child
-                if parallel {
-                    indices.push(dep_index);
-
+            for dep_task in self
+                .internal_resolve_tasks_from_target(&dep.target, true)
+                .await?
+            {
+                if let Some(dep_index) = self.internal_run_task(&dep_task, reqs, Some(dep)).await? {
+                    // When parallel, parent depends on child
+                    if parallel {
+                        indexes.push(Some(dep_index));
+                    }
                     // When serial, next child depends on previous child
-                } else if let Some(prev) = previous_target_index {
-                    self.link_requirements(dep_index, vec![prev]);
-                }
+                    else {
+                        self.link_requirements(dep_index, vec![previous_target_index]);
+                    }
 
-                previous_target_index = Some(dep_index);
+                    previous_target_index = Some(dep_index);
+                }
             }
         }
 
         if !parallel {
-            if let Some(index) = previous_target_index {
-                indices.push(index);
-            }
+            indexes.push(previous_target_index);
         }
 
-        Ok(indices)
+        Ok(indexes)
     }
 
-    #[instrument(skip_all)]
-    pub fn run_task_dependents(
+    pub async fn run_task_dependents(
         &mut self,
         task: &Task,
         reqs: &RunRequirements,
-    ) -> miette::Result<Vec<NodeIndex>> {
-        let mut indices = vec![];
+    ) -> miette::Result<Vec<Option<NodeIndex>>> {
+        let mut indexes = vec![];
 
         for dep_target in self.workspace_graph.tasks.dependents_of(task) {
-            let (_, dep_indices) = self.run_task_by_target(dep_target, reqs)?;
-
-            for dep_index in dep_indices {
-                indices.push(dep_index);
+            for dep_task in self
+                .internal_resolve_tasks_from_target(&dep_target, true)
+                .await?
+            {
+                indexes.push(self.internal_run_task(&dep_task, reqs, None).await?);
             }
         }
 
-        Ok(indices)
+        Ok(indexes)
     }
 
-    pub fn run_task_by_target<T: AsRef<Target>>(
+    #[instrument(skip_all)]
+    async fn internal_resolve_tasks_from_target(
         &mut self,
-        target: T,
-        reqs: &RunRequirements,
-    ) -> miette::Result<(FxHashSet<Target>, FxHashSet<NodeIndex>)> {
-        self.run_task_by_target_with_config(target, reqs, None)
-    }
-
-    pub fn run_task_by_target_with_config<T: AsRef<Target>>(
-        &mut self,
-        target: T,
-        reqs: &RunRequirements,
-        config: Option<&TaskDependencyConfig>,
-    ) -> miette::Result<(FxHashSet<Target>, FxHashSet<NodeIndex>)> {
-        let target = target.as_ref();
-        let mut inserted_targets = FxHashSet::default();
-        let mut inserted_indices = FxHashSet::default();
+        target: &Target,
+        allow_internal: bool,
+    ) -> miette::Result<Vec<Arc<Task>>> {
+        let mut tasks = vec![];
 
         match &target.scope {
             // :task
@@ -577,16 +612,11 @@ impl<'app> ActionGraphBuilder<'app> {
                         .workspace_graph
                         .get_task_from_project(&project.id, &target.task_id)
                     {
-                        if task.is_internal() {
+                        if !allow_internal && task.is_internal() {
                             continue;
                         }
 
-                        if let Some(index) =
-                            self.run_task_with_config(&project, &task, reqs, config)?
-                        {
-                            inserted_targets.insert(task.target.clone());
-                            inserted_indices.insert(index);
-                        }
+                        tasks.push(task);
                     }
                 }
             }
@@ -595,21 +625,17 @@ impl<'app> ActionGraphBuilder<'app> {
                 return Err(TargetError::NoDepsInRunContext.into());
             }
             // project:task
-            TargetScope::Project(project_locator) => {
-                let project = self.workspace_graph.get_project(project_locator)?;
+            TargetScope::Project(project_id) => {
                 let task = self
                     .workspace_graph
-                    .get_task_from_project(&project.id, &target.task_id)?;
+                    .get_task_from_project(project_id, &target.task_id)?;
 
                 // Don't allow internal tasks to be ran
-                if task.is_internal() && reqs.has_target(&task.target) {
+                if !allow_internal && task.is_internal() {
                     return Err(TaskGraphError::UnconfiguredTarget(task.target.clone()).into());
                 }
 
-                if let Some(index) = self.run_task_with_config(&project, &task, reqs, config)? {
-                    inserted_targets.insert(task.target.to_owned());
-                    inserted_indices.insert(index);
-                }
+                tasks.push(task);
             }
             // #tag:task
             TargetScope::Tag(tag) => {
@@ -623,16 +649,11 @@ impl<'app> ActionGraphBuilder<'app> {
                         .workspace_graph
                         .get_task_from_project(&project.id, &target.task_id)
                     {
-                        if task.is_internal() {
+                        if !allow_internal && task.is_internal() {
                             continue;
                         }
 
-                        if let Some(index) =
-                            self.run_task_with_config(&project, &task, reqs, config)?
-                        {
-                            inserted_targets.insert(task.target.clone());
-                            inserted_indices.insert(index);
-                        }
+                        tasks.push(task);
                     }
                 }
             }
@@ -642,207 +663,295 @@ impl<'app> ActionGraphBuilder<'app> {
             }
         };
 
-        Ok((inserted_targets, inserted_indices))
+        Ok(tasks)
     }
 
     #[instrument(skip_all)]
-    pub fn run_from_requirements(
+    async fn internal_resolve_tasks_from_target_locator(
         &mut self,
-        reqs: RunRequirements,
-    ) -> miette::Result<Vec<NodeIndex>> {
-        let mut inserted_nodes = vec![];
-        let mut initial_targets = vec![];
+        locator: &TargetLocator,
+        allow_internal: bool,
+    ) -> miette::Result<Vec<Arc<Task>>> {
+        let mut tasks = vec![];
 
-        if let Some(affected) = &mut self.affected {
-            affected.set_ci_check(reqs.ci_check);
-        }
+        match locator {
+            TargetLocator::GlobMatch {
+                project_glob,
+                task_glob,
+                scope,
+                ..
+            } => {
+                let mut is_all = false;
+                let mut do_query = false;
+                let mut projects = vec![];
 
-        // Track the qualified as an initial target
-        for locator in reqs.target_locators.clone() {
-            match locator {
-                TargetLocator::GlobMatch {
-                    project_glob,
-                    task_glob,
-                    scope,
-                    ..
-                } => {
-                    let mut is_all = false;
-                    let mut do_query = false;
-                    let mut projects = vec![];
-
-                    // Query for all applicable projects first since we can't
-                    // query projects + tasks at the same time
-                    if let Some(glob) = project_glob {
-                        let query = if let Some(tag_glob) = glob.strip_prefix('#') {
-                            format!("tag~{tag_glob}")
-                        } else {
-                            format!("project~{glob}")
-                        };
-
-                        projects = self.workspace_graph.query_projects(build_query(&query)?)?;
-                        do_query = !projects.is_empty();
+                // Query for all applicable projects first since we can't
+                // query projects + tasks at the same time
+                if let Some(glob) = project_glob {
+                    let query = if let Some(tag_glob) = glob.strip_prefix('#') {
+                        format!("tag~{tag_glob}")
                     } else {
-                        match scope {
-                            Some(TargetScope::All) => {
-                                is_all = true;
-                                do_query = true;
-                            }
-                            _ => {
-                                // Don't query for the other scopes,
-                                // since they're not valid from the run context
-                            }
-                        };
-                    }
+                        format!("project~{glob}")
+                    };
 
-                    // Then query for all tasks within the queried projects
-                    if do_query {
-                        let mut query = format!("task~{task_glob}");
-
-                        if !is_all {
-                            query = format!(
-                                "project=[{}] && {query}",
-                                projects
-                                    .into_iter()
-                                    .map(|project| project.id.to_string())
-                                    .collect::<Vec<_>>()
-                                    .join(",")
-                            );
+                    projects = self.workspace_graph.query_projects(build_query(&query)?)?;
+                    do_query = !projects.is_empty();
+                } else {
+                    match scope {
+                        Some(TargetScope::All) => {
+                            is_all = true;
+                            do_query = true;
                         }
+                        _ => {
+                            // Don't query for the other scopes,
+                            // since they're not valid from the run context
+                        }
+                    };
+                }
 
-                        let tasks = self.workspace_graph.query_tasks(build_query(&query)?)?;
+                // Then query for all tasks within the queried projects
+                if do_query {
+                    let mut query = format!("task~{task_glob}");
 
-                        initial_targets.extend(
-                            tasks
+                    if !is_all {
+                        query = format!(
+                            "project=[{}] && {query}",
+                            projects
                                 .into_iter()
-                                .map(|task| task.target.clone())
-                                .collect::<Vec<_>>(),
+                                .map(|project| project.id.to_string())
+                                .collect::<Vec<_>>()
+                                .join(",")
                         );
                     }
-                }
-                TargetLocator::Qualified(target) => {
-                    if target.scope == TargetScope::OwnSelf {
-                        initial_targets.push(Target::new(
-                            &self.workspace_graph.get_project_from_path(None)?.id,
-                            target.task_id,
-                        )?);
-                    } else {
-                        initial_targets.push(target);
+
+                    for task in self.workspace_graph.query_tasks(build_query(&query)?)? {
+                        if !allow_internal && task.is_internal() {
+                            continue;
+                        }
+
+                        tasks.push(task);
                     }
                 }
-                TargetLocator::TaskFromWorkingDir(task_id) => {
-                    initial_targets.push(Target::new(
+            }
+            TargetLocator::Qualified(target) => {
+                let target = if target.scope == TargetScope::OwnSelf {
+                    Target::new(
                         &self.workspace_graph.get_project_from_path(None)?.id,
-                        task_id,
-                    )?);
-                }
-            };
-        }
+                        &target.task_id,
+                    )?
+                } else {
+                    target.to_owned()
+                };
 
-        // Determine affected tasks before building
+                tasks.extend(
+                    self.internal_resolve_tasks_from_target(&target, allow_internal)
+                        .await?,
+                );
+            }
+            TargetLocator::TaskFromWorkingDir(task_id) => {
+                let target = Target::new(
+                    &self.workspace_graph.get_project_from_path(None)?.id,
+                    task_id,
+                )?;
+
+                tasks.extend(
+                    self.internal_resolve_tasks_from_target(&target, allow_internal)
+                        .await?,
+                );
+            }
+        };
+
+        Ok(tasks)
+    }
+
+    #[instrument(skip_all)]
+    async fn internal_run_task(
+        &mut self,
+        task: &Task,
+        reqs: &RunRequirements,
+        config: Option<&TaskDependencyConfig>,
+    ) -> miette::Result<Option<NodeIndex>> {
+        let project = self
+            .workspace_graph
+            .get_project(task.target.get_project_id().unwrap())?;
+
+        let child_reqs = RunRequirements {
+            ci: reqs.ci,
+            ci_check: reqs.ci_check,
+            dependents: false,
+            interactive: reqs.interactive,
+        };
+
+        // Abort early if not affected
         if let Some(affected) = &mut self.affected {
-            affected.track_tasks()?;
+            if !affected.is_task_marked(task) {
+                return Ok(None);
+            }
         }
 
-        // Then build and track initial and primary
-        for target in initial_targets {
-            let (inserted_targets, inserted_indices) = self.run_task_by_target(&target, &reqs)?;
+        // These tasks shouldn't actually run, so filter them out
+        if self.passthrough_targets.contains(&task.target) {
+            trace!(
+                task_target = task.target.as_str(),
+                "Not running task {} because it has been marked as passthrough",
+                color::id(&task.target.id),
+            );
 
-            self.initial_targets.insert(target.clone());
-            self.primary_targets.extend(inserted_targets);
-
-            inserted_nodes.extend(inserted_indices);
+            return Ok(None);
         }
 
-        Ok(inserted_nodes)
+        // Only apply checks when requested. This applies to `moon ci`,
+        // but not `moon run`, since the latter should be able to
+        // manually run local tasks in CI (deploys, etc).
+        if reqs.ci && reqs.ci_check && !task.should_run_in_ci() {
+            self.passthrough_targets.insert(task.target.clone());
+
+            debug!(
+                task_target = task.target.as_str(),
+                "Not running task {} because {} is false",
+                color::id(&task.target.id),
+                color::property("runInCI"),
+            );
+
+            // Dependents may still want to run though!
+            if reqs.dependents {
+                Box::pin(self.run_task_dependents(task, &child_reqs)).await?;
+            }
+
+            return Ok(None);
+        }
+
+        // Create the node
+        let mut args = vec![];
+        let mut env = FxHashMap::default();
+
+        if let Some(config) = config {
+            args.extend(parse_task_args(&config.args)?);
+            env.extend(config.env.clone());
+        }
+
+        let node = ActionNode::run_task(RunTaskNode {
+            args,
+            env,
+            interactive: task.is_interactive() || reqs.interactive,
+            persistent: task.is_persistent(),
+            priority: task.options.priority.get_level(),
+            target: task.target.to_owned(),
+            id: None,
+        });
+
+        // Check if the node exists to avoid all the overhead below
+        if let Some(index) = self.get_index_from_node(&node) {
+            return Ok(Some(index));
+        }
+
+        // Create initial edges
+        let mut edges = vec![self.sync_project(&project).await?];
+
+        edges.extend(
+            self.install_dependencies_by_toolchains(&project, &task.toolchains)
+                .await?,
+        );
+
+        // Insert and then link edges
+        let index = self.insert_node(node);
+
+        if !task.deps.is_empty() {
+            edges.extend(Box::pin(self.run_task_dependencies(task, &child_reqs)).await?);
+        }
+
+        self.link_requirements(index, edges);
+
+        // And possibly dependents
+        if reqs.dependents {
+            Box::pin(self.run_task_dependents(task, &child_reqs)).await?;
+        }
+
+        Ok(Some(index))
     }
 
     #[instrument(skip_all)]
-    pub fn setup_toolchain(&mut self, runtime: &Runtime) -> Option<NodeIndex> {
+    pub async fn setup_toolchain_legacy(
+        &mut self,
+        runtime: &Runtime,
+    ) -> miette::Result<Option<NodeIndex>> {
+        let sync_workspace_index = self.sync_workspace().await;
+
+        // Explicitly disabled
         if !self.options.setup_toolchains.is_enabled(&runtime.toolchain) || runtime.is_system() {
-            return None;
+            return Ok(sync_workspace_index);
         }
 
-        let node = ActionNode::setup_toolchain_legacy(SetupToolchainLegacyNode {
-            runtime: runtime.to_owned(),
-        });
+        let index = self.insert_node(ActionNode::setup_toolchain_legacy(
+            SetupToolchainLegacyNode {
+                runtime: runtime.to_owned(),
+            },
+        ));
 
-        if let Some(index) = self.get_index_from_node(&node) {
-            return Some(*index);
-        }
+        self.link_requirements(index, vec![sync_workspace_index]);
 
-        let sync_workspace_index = self.sync_workspace();
-        let index = self.insert_node(node);
-
-        if let Some(edge) = sync_workspace_index {
-            self.link_requirements(index, vec![edge]);
-        }
-
-        Some(index)
+        Ok(Some(index))
     }
 
     #[instrument(skip_all)]
-    pub fn setup_toolchain_plugin(&mut self, spec: &ToolchainSpec) -> Option<NodeIndex> {
+    pub async fn setup_toolchain(
+        &mut self,
+        spec: &ToolchainSpec,
+    ) -> miette::Result<Option<NodeIndex>> {
+        let sync_workspace_index = self.sync_workspace().await;
+
+        // Explicitly disabled
         if !self.options.setup_toolchains.is_enabled(&spec.id) || spec.is_system() {
-            return None;
+            return Ok(sync_workspace_index);
         }
 
-        let node = ActionNode::setup_toolchain(SetupToolchainNode {
+        let toolchain = self.app_context.toolchain_registry.load(&spec.id).await?;
+
+        // Toolchain does not support tier 3
+        if !toolchain.supports_tier_3().await {
+            return Ok(sync_workspace_index);
+        }
+
+        let index = self.insert_node(ActionNode::setup_toolchain(SetupToolchainNode {
             spec: spec.to_owned(),
-        });
+        }));
 
-        if let Some(index) = self.get_index_from_node(&node) {
-            return Some(*index);
-        }
+        self.link_requirements(index, vec![sync_workspace_index]);
 
-        let sync_workspace_index = self.sync_workspace();
-        let index = self.insert_node(node);
-
-        if let Some(edge) = sync_workspace_index {
-            self.link_requirements(index, vec![edge]);
-        }
-
-        Some(index)
+        Ok(Some(index))
     }
 
     #[instrument(skip_all)]
-    pub fn sync_project(&mut self, project: &Project) -> miette::Result<Option<NodeIndex>> {
+    pub async fn sync_project(&mut self, project: &Project) -> miette::Result<Option<NodeIndex>> {
         self.internal_sync_project(project, &mut FxHashSet::default())
+            .await
     }
 
-    fn internal_sync_project(
+    async fn internal_sync_project(
         &mut self,
         project: &Project,
         cycle: &mut FxHashSet<Id>,
     ) -> miette::Result<Option<NodeIndex>> {
+        let sync_workspace_index = self.sync_workspace().await;
+
+        // Explicitly disabled
         if !self.options.sync_projects.is_enabled(&project.id) {
-            return Ok(None);
+            return Ok(sync_workspace_index);
         }
 
-        let node = ActionNode::sync_project(SyncProjectNode {
-            project_id: project.id.clone(),
-        });
-
-        if let Some(index) = self.get_index_from_node(&node) {
-            return Ok(Some(*index));
+        // Abort early if not affected
+        if let Some(affected) = &mut self.affected {
+            if !affected.is_project_marked(project) {
+                return Ok(None);
+            }
         }
 
         cycle.insert(project.id.clone());
 
-        // Determine affected state
-        if let Some(affected) = &mut self.affected {
-            if let Some(by) = affected.is_project_affected(project) {
-                affected.mark_project_affected(project, by)?;
-            }
-        }
-
-        // Syncing requires the language's tool to be installed
-        let mut edges = vec![];
-
-        if let Some(edge) = self.sync_workspace() {
-            edges.push(edge);
-        }
-
-        let index = self.insert_node(node);
+        let mut edges = vec![sync_workspace_index];
+        let index = self.insert_node(ActionNode::sync_project(SyncProjectNode {
+            project_id: project.id.clone(),
+        }));
 
         // And we should also depend on other projects
         if self.options.sync_project_dependencies {
@@ -853,9 +962,11 @@ impl<'app> ActionGraphBuilder<'app> {
 
                 let dep_project = self.workspace_graph.get_project(&dep_project_id)?;
 
-                if let Some(dep_project_index) = self.internal_sync_project(&dep_project, cycle)? {
+                if let Some(dep_project_index) =
+                    Box::pin(self.internal_sync_project(&dep_project, cycle)).await?
+                {
                     if index != dep_project_index {
-                        edges.push(dep_project_index);
+                        edges.push(Some(dep_project_index));
                     }
                 }
             }
@@ -868,26 +979,33 @@ impl<'app> ActionGraphBuilder<'app> {
         Ok(Some(index))
     }
 
-    pub fn sync_workspace(&mut self) -> Option<NodeIndex> {
+    pub async fn sync_workspace(&mut self) -> Option<NodeIndex> {
         if !self.options.sync_workspace {
             return None;
         }
 
-        let node = ActionNode::sync_workspace();
-
-        if let Some(index) = self.get_index_from_node(&node) {
-            return Some(*index);
-        }
-
-        Some(self.insert_node(node))
+        Some(self.insert_node(ActionNode::sync_workspace()))
     }
 
     // PRIVATE
 
-    fn link_requirements(&mut self, index: NodeIndex, edges: Vec<NodeIndex>) {
+    fn get_index_from_node(&self, node: &ActionNode) -> Option<NodeIndex> {
+        self.graph
+            .node_references()
+            .find(|(_, n)| *n == node)
+            .map(|(i, _)| i)
+    }
+
+    fn link_requirements(&mut self, index: NodeIndex, edges: Vec<Option<NodeIndex>>) {
+        let edges = edges.into_iter().flatten().collect::<Vec<_>>();
+
+        if edges.is_empty() {
+            return;
+        }
+
         trace!(
             index = index.index(),
-            requires = ?edges.iter().map(|i| i.index()).collect::<Vec<_>>(),
+            requires = ?edges.iter().map(|edge| edge.index()).collect::<Vec<_>>(),
             "Linking requirements for index"
         );
 
@@ -899,15 +1017,18 @@ impl<'app> ActionGraphBuilder<'app> {
     }
 
     fn insert_node(&mut self, node: ActionNode) -> NodeIndex {
-        let index = self.graph.add_node(node.clone());
+        if let Some(index) = self.get_index_from_node(&node) {
+            return index;
+        }
+
+        let label = node.label();
+        let index = self.graph.add_node(node);
 
         debug!(
             index = index.index(),
             "Adding {} to graph",
-            color::muted_light(node.label())
+            color::muted_light(label)
         );
-
-        self.indices.insert(node, index);
 
         index
     }
