+++ crates/action-graph/src/lib.rs
@@ -1,6 +1,5 @@
 mod action_graph;
 mod action_graph_builder;
-pub mod action_graph_builder2;
 mod action_graph_error;
 
 pub use action_graph::*;
