+++ crates/app/src/commands/sync.rs
@@ -42,18 +42,15 @@ pub async fn sync(session: MoonSession) -> AppResult {
         .build_action_graph_with_options(ActionGraphBuilderOptions::default())
         .await?;
 
-    action_graph_builder.sync_workspace();
+    action_graph_builder.sync_workspace().await;
 
     for project in workspace_graph.projects.get_all_unexpanded() {
-        action_graph_builder.sync_project(project)?;
+        action_graph_builder.sync_project(project).await?;
     }
 
-    run_action_pipeline(
-        &session,
-        action_graph_builder.build_context(),
-        action_graph_builder.build(),
-    )
-    .await?;
+    let (action_context, action_graph) = action_graph_builder.build();
+
+    run_action_pipeline(&session, action_context, action_graph).await?;
 
     session.console.render(element! {
         Container {
