+++ crates/app/src/session.rs
@@ -98,6 +98,7 @@ impl MoonSession {
             workspace_graph,
             ActionGraphBuilderOptions {
                 install_dependencies: config.install_dependencies.clone(),
+                setup_environment: true.into(),
                 setup_toolchains: true.into(),
                 sync_projects: config.sync_projects.clone(),
                 sync_project_dependencies: config.sync_project_dependencies,
