+++ crates/test-utils/src/lib.rs
@@ -1,10 +1,8 @@
 mod platform_manager;
 mod sandbox;
-mod workspace_graph;
 mod workspace_mocker;
 
 pub use platform_manager::*;
 pub use sandbox::*;
 pub use starbase_sandbox::{predicates, pretty_assertions};
-pub use workspace_graph::*;
 pub use workspace_mocker::*;
